# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/cities": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a City",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Create a City",
                "parameters": [
                    {
                        "description": "Create a City",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_dto.CreateUpdateCityRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "City response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_dto.CityResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/cities/get-by-filter": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get Cities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Get Cities",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_dto.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_dto.PageList-github_com_mohammad-ebrahimi-it_car-shoping_api_dto_CityResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/cities/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get a City",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Get a City",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_dto.CityResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update a City",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Update a City",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update a City",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_dto.CreateUpdateCityRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "City response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_dto.CityResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete a City",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Delete a City",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_helper.BaseHttpResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_mohammad-ebrahimi-it_car-shoping_api_dto.CityResponse": {
            "type": "object",
            "properties": {
                "country": {
                    "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_dto.CountryResponse"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_mohammad-ebrahimi-it_car-shoping_api_dto.CountryResponse": {
            "type": "object",
            "properties": {
                "cities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_dto.CityResponse"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_mohammad-ebrahimi-it_car-shoping_api_dto.CreateUpdateCityRequest": {
            "type": "object",
            "required": [
                "countryId",
                "name"
            ],
            "properties": {
                "countryId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                }
            }
        },
...
        "github_com_mohammad-ebrahimi-it_car-shoping_api_dto.Filter": {
            "type": "object",
            "properties": {
                "filter_type": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_mohammad-ebrahimi-it_car-shoping_api_dto.PageList-github_com_mohammad-ebrahimi-it_car-shoping_api_dto_CityResponse": {
            "type": "object",
            "properties": {
                "has_next_page": {
                    "type": "boolean"
                },
                "has_previous_page": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_dto.CityResponse"
                    }
                },
                "page_number": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_mohammad-ebrahimi-it_car-shoping_api_dto.PaginationInputWithFilter": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_dto.Filter"
                    }
                },
                "page_number": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "sort": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_dto.Sort"
                    }
                }
            }
        },
...
        "github_com_mohammad-ebrahimi-it_car-shoping_api_dto.Sort": {
            "type": "object",
            "properties": {
                "col_id": {
                    "type": "string"
                },
                "sort": {
                    "type": "string"
                }
            }
        },
        "github_com_mohammad-ebrahimi-it_car-shoping_api_helper.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "result": {},
                "resultCode": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "validationErrors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_mohammad-ebrahimi-it_car-shoping_api_validations.ValidationError"
                    }
                }
            }
        },
        "github_com_mohammad-ebrahimi-it_car-shoping_api_validations.ValidationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AuthBearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "bearer"
        }
    }
}